requirements.txt
asgiref==3.3.1
Django==3.1.6
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
flake8==3.8.4
install==1.3.4
mccabe==0.6.1
phonenumbers==8.12.18
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2021.1
sqlparse==0.4.1
psycopg2==2.8.6
amqp==5.0.5
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
factory-boy==3.2.0
Faker==6.1.1
kombu==5.0.2
prompt-toolkit==3.0.16
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
six==1.15.0
text-unidecode==1.3
vine==5.0.0
wcwidth==0.2.5